feat(network): move stories to the dedicated storybook workspace
